pyproject.toml
[project]
name = "sendig-extractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "camelot-py>=0.11.0",
    "django>=5.2.1",
    "numpy==1.26.0",
    "opencv-python>=4.11.0.86",
    "pdfminer-six>=20250506",
    "pycryptodome>=3.22.0",
    "pymupdf>=1.25.5",
    "pypdf2>=3.0.1",
    "pyreadstat>=1.2.8",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "tqdm>=4.67.1",
    "whitenoise>=6.9.0",
    "xport>=3.6.1",
]
